Solved signal strength issue
As far as MBM module cannot provide LTE signal strength it was disabled.
In the original implementation it was initialized with 0x7FFFFFFF. But to
disable displaying fake values it should be changed to "-1". So, it was
done.

Also were added debug notification on preferred network change and a bit
reworked network type selecting mechanism to avoid situation with
unsupported network type.

Signed-off-by: Sergey Shcherbakov <shchers@gmail.com>